fix: 상품검색 api 수정 (전체목록불러오고 프론트에서필터링)
